pyproject.toml
[project]
name = "chandra-project"
version = "0.1.0"
description = "This project aims to develop a modular attachment for robots to assist visually impaired individuals by enhancing their ability to navigate and interact with their surroundings."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "loguru>=0.7.3",
    "openai-whisper>=20240930",
    "pre-commit>=4.1.0",
    "smolagents>=1.10.0",
    "soundfile>=0.13.1",
    "streamlit>=1.43.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "word2number>=1.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
